refactored plot interface moved view model to separate file packaged visual
