+++ src/constants.ts
@@ -107,8 +107,7 @@ export class ArrayConstants {
         singlehue: ['Blues', 'Greens', 'Greys', 'Oranges', 'Purples', 'Reds'],
         diverging: ['BrBG', 'PiYG', 'PRGn', 'PuOr', 'RdBu', 'RdGy', 'RdYlBu', 'RdYlGn', 'Spectral'],
     };
-    // static readonly rolloutColors = ['#ffffff', '#a8a8a8', '#222222', '#4daf4a', '#ff0000'];
-    // static readonly rolloutColors = ['#ffffff', '#2ca25f', '#de2d26'];
+
     static readonly legendColors = {
         OZE: '#e41a1c',
         GZE: '#377eb8',
@@ -144,8 +143,6 @@ export class ArrayConstants {
         '#e5c494',
         '#b3b3b3',
     ];
-
-    //static readonly rolloutNames = ['nicht gewalzt', 'BBS gewalzt', 'Beize gewalzt', 'KB gewalzt ohne Fehler', 'KB gewalzt mit Fehler'];
 }
 
 export class NumberConstants {
